split tests into popup and overlay
