+++ cypress/.gitignore
@@ -1,3 +1,4 @@
 public/playkit-call-to-action.js
 public/playkit-call-to-action.js.map
 screenshots
+downloads
