+++ cypress/e2e/call-to-action.overlay.cy.ts
@@ -1,4 +1,14 @@
-import {expectCloseButton, expectContains, expectLoadMedia, expectWindowOpen, getOverlayElement} from './utils/env';
+import {
+  expectCloseButton,
+  expectContains,
+  expectElementDoesntExist,
+  expectElementDoesntExistAfter,
+  expectElementExists,
+  expectElementExistsAt,
+  expectLoadMedia,
+  expectWindowOpen,
+  getOverlayElement
+} from './utils/env';
 
 const TITLE = 'cta title';
 const DESCRIPTION = 'cta description';
@@ -13,7 +23,7 @@ const expectContainsInOverlay = (pluginConfig: any, texts: string[]) => {
   expectContains(pluginConfig, texts, getOverlayElement);
 };
 const expectCloseButtonInOverlay = (pluginConfig: any) => {
-  expectCloseButton(pluginConfig, getOverlayElement, getCloseButton);
+  expectCloseButton(pluginConfig, getCloseButton, getOverlayElement);
 };
 const expectWindowOpenInOverlay = (pluginConfig: any, buttonLabel: string, buttonLink: string) => {
   expectWindowOpen(pluginConfig, buttonLabel, buttonLink, getOverlayElement);
@@ -21,198 +31,439 @@ const expectWindowOpenInOverlay = (pluginConfig: any, buttonLabel: string, butto
 const expectLoadMediaInOverlay = (pluginConfig: any, buttonLabel: string, buttonLink: string) => {
   expectLoadMedia(pluginConfig, buttonLabel, buttonLink, getOverlayElement);
 };
+const expectOverlayExists = (pluginConfig: object) => {
+  expectElementExists(pluginConfig, getOverlayElement);
+};
+const expectOverlayDoesntExist = (pluginConfig: object) => {
+  expectElementDoesntExist(pluginConfig, getOverlayElement);
+};
+const expectOverlayExistsAt = (pluginConfig: object, expectedStartTime: number) => {
+  expectElementExistsAt(pluginConfig, expectedStartTime, getOverlayElement);
+};
+const expectOverlayDoesntExistAfter = (pluginConfig: object, expectedStartTime: number, expectedDuration: number) => {
+  expectElementDoesntExistAfter(pluginConfig, expectedStartTime, expectedDuration, getOverlayElement);
+};
 
 describe('call to action overlay', () => {
-  it('should show title', () => {
-    expectContainsInOverlay({messages: [{title: TITLE, timing: {showOnStart: true}}]}, [TITLE]);
+  describe('message validation', () => {
+    describe('message content', () => {
+      it('should not show message if title, description and buttons are not set', () => {
+        expectOverlayDoesntExist({messages: [{timing: {showOnStart: true}}]});
+      });
+      describe('title', () => {
+        it('should show message if title is set', () => {
+          expectOverlayExists({messages: [{title: 'aaa', timing: {showOnStart: true}}]});
+        });
+      });
+      describe('description', () => {
+        it('should show message if description is set', () => {
+          expectOverlayExists({messages: [{description: 'aaa', timing: {showOnStart: true}}]});
+        });
+      });
+      describe('buttons', () => {
+        it('should show message with one button', () => {
+          expectOverlayExists({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {showOnStart: true}}]});
+        });
+        it('should show message with two buttons', () => {
+          expectOverlayExists({
+            messages: [
+              {
+                buttons: [
+                  {label: 'aaa', link: 'aaa'},
+                  {label: 'bbb', link: 'bbb'}
+                ],
+                timing: {showOnStart: true}
+              }
+            ]
+          });
+        });
+        it('should not show message if button label is not set', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{link: 'aaa'}], timing: {showOnStart: true}}]});
+        });
+        it('should not show message if button label is not a string', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{label: -1, link: 'aaa'}], timing: {showOnStart: true}}]});
+        });
+        it('should not show message if button label is an empty string', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{label: '', link: 'aaa'}], timing: {showOnStart: true}}]});
+        });
+        it('should not show message if button link is not set', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{label: ''}], timing: {showOnStart: true}}]});
+        });
+        it('should not show message if button link is not a string', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{label: 'aaa', link: -1}], timing: {showOnStart: true}}]});
+        });
+        it('should not show message if button link is an empty string', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{label: 'aaa', link: ''}], timing: {showOnStart: true}}]});
+        });
+      });
+    });
+    describe('message timing', () => {
+      it('should not show message if timing is undefined', () => {
+        expectOverlayDoesntExist({messages: [{buttons: [{label: 'aaa', link: 'aaa'}]}]});
+      });
+      it('should not show message if none of the timing fields are set', () => {
+        expectOverlayDoesntExist({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {}}]});
+      });
+      describe('showOnStart', () => {
+        it('should show message if showOnStart is true', () => {
+          expectOverlayExists({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {showOnStart: true}}]});
+        });
+        it('should not show message if showOnStart is not true', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {showOnStart: false}}]});
+        });
+        it('should not show message if showOnStart is not boolean', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {showOnStart: 1}}]});
+        });
+      });
+      describe('showOnEnd', () => {
+        it('should show message if showOnEnd is true', () => {
+          expectOverlayExists({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {showOnEnd: true}}]});
+        });
+        it('should not show message if showOnEnd is not true', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {showOnEnd: false}}]});
+        });
+        it('should not show message if showOnEnd is not boolean', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {showOnEnd: 1}}]});
+        });
+      });
+      describe('timeFromStart', () => {
+        it('should show message if timeFromStart is a positive number', () => {
+          expectOverlayExists({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {timeFromStart: 1}}]});
+        });
+        it('should show message if timeFromStart is zero', () => {
+          expectOverlayExists({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {timeFromStart: 0}}]});
+        });
+        it('should not show message if timeFromStart is a negative number', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {timeFromStart: -1}}]});
+        });
+        it('should not show message if timeFromStart is not a number', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {timeFromStart: 'aaa'}}]});
+        });
+      });
+      describe('timeFromEnd', () => {
+        it('should show message if timeFromEnd is a positive number', () => {
+          expectOverlayExists({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {timeFromEnd: 1}}]});
+        });
+        it('should show message if timeFromEnd is zero', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {timeFromEnd: 0}}]});
+        });
+        it('should not show message if timeFromEnd is not a negative number', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {timeFromEnd: -1}}]});
+        });
+        it('should not show message if timeFromEnd is not a number', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {timeFromEnd: 'aaa'}}]});
+        });
+      });
+      describe('duration', () => {
+        it('should show message if duration is not a positive number', () => {
+          expectOverlayExists({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {showOnStart: true, duration: 1}}]});
+        });
+        it('should not show message if duration is zero', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {showOnStart: true, duration: 0}}]});
+        });
+        it('should not show message if duration is negative', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {showOnStart: true, duration: -1}}]});
+        });
+        it('should not show message if duration is not a number', () => {
+          expectOverlayDoesntExist({messages: [{buttons: [{label: 'aaa', link: 'aaa'}], timing: {showOnStart: true, duration: 'aaa'}}]});
+        });
+      });
+    });
   });
-  it('should show description', () => {
-    expectContainsInOverlay({messages: [{description: DESCRIPTION, timing: {showOnStart: true}}]}, [DESCRIPTION]);
+  describe('message timing', () => {
+    describe('start time and duration', () => {
+      describe('show on start', () => {
+        it('should display message on playback start', () => {
+          expectOverlayExistsAt(
+            {
+              messages: [
+                {
+                  title: TITLE,
+                  timing: {showOnStart: true}
+                }
+              ]
+            },
+            0
+          );
+        });
+        it('should hide message after duration', () => {
+          expectOverlayDoesntExistAfter(
+            {
+              messages: [
+                {
+                  title: TITLE,
+                  timing: {showOnStart: true, duration: 2}
+                }
+              ]
+            },
+            0,
+            2
+          );
+        });
+      });
+      describe('show on end', () => {
+        it('should display message on playback end', () => {
+          expectOverlayExistsAt(
+            {
+              messages: [
+                {
+                  title: TITLE,
+                  timing: {showOnEnd: true}
+                }
+              ]
+            },
+            4
+          );
+        });
+        it('should hide message after duration', () => {
+          expectOverlayDoesntExistAfter(
+            {
+              messages: [
+                {
+                  title: TITLE,
+                  timing: {showOnEnd: true, duration: 2}
+                }
+              ]
+            },
+            4,
+            2
+          );
+        });
+      });
+      describe('time from start', () => {
+        it('should display message on time from start', () => {
+          expectOverlayExistsAt(
+            {
+              messages: [
+                {
+                  title: TITLE,
+                  timing: {timeFromStart: 1}
+                }
+              ]
+            },
+            1
+          );
+        });
+        it('should hide message after duration', () => {
+          expectOverlayDoesntExistAfter(
+            {
+              messages: [
+                {
+                  title: TITLE,
+                  timing: {timeFromStart: 1, duration: 2}
+                }
+              ]
+            },
+            1,
+            2
+          );
+        });
+      });
+      describe('time from end', () => {
+        it('should display message on time from end', () => {
+          expectOverlayExistsAt(
+            {
+              messages: [
+                {
+                  title: TITLE,
+                  timing: {timeFromEnd: 2}
+                }
+              ]
+            },
+            2
+          );
+        });
+        it('should hide message after duration', () => {
+          expectOverlayDoesntExistAfter(
+            {
+              messages: [
+                {
+                  title: TITLE,
+                  timing: {timeFromEnd: 3, duration: 2}
+                }
+              ]
+            },
+            1,
+            2
+          );
+        });
+      });
+    });
+    //describe('seeked event');
+    // message ended normally
+    // message was closed and redisplay is false
+    // message was closed and redisplay is true
+    // seek into a message range
+    // message has duration
+    // message has no duration
+    // describe('multiple messages', () => {
+    //   describe('one message on start, the other at time from start');
+    //   describe('one message on start, the other at time from end');
+    //   describe('one message on start, the other on end');
+    //   describe('two messages at time from start');
+    //   describe('one at time from start, the other at time from end');
+    //   describe('two messages at time from end');
+    //   describe('one message on end, the other at time from start');
+    //   describe('one message on end, the other at time from end');
+    // });
   });
-  it('should show one button', () => {
-    expectContainsInOverlay(
-      {
-        messages: [{buttons: [{label: BUTTON_1_LABEL, link: BUTTON_LINK_URL}], timing: {showOnStart: true}}]
-      },
-      [BUTTON_1_LABEL]
-    );
-  });
-  it('should show two buttons', () => {
-    expectContainsInOverlay(
-      {
-        messages: [
-          {
-            buttons: [
-              {label: BUTTON_1_LABEL, link: BUTTON_LINK_URL},
-              {label: BUTTON_2_LABEL, link: BUTTON_LINK_ENTRY}
-            ],
-            timing: {showOnStart: true}
-          }
-        ]
-      },
-      [BUTTON_1_LABEL, BUTTON_2_LABEL]
-    );
-  });
-  it('should show title and description', () => {
-    expectContainsInOverlay(
-      {
-        messages: [{title: TITLE, description: DESCRIPTION, timing: {showOnStart: true}}]
-      },
-      [TITLE, DESCRIPTION]
-    );
-  });
-  it('should show title and one button', () => {
-    expectContainsInOverlay(
-      {
-        messages: [
-          {
-            title: TITLE,
-            buttons: [{label: BUTTON_1_LABEL, link: BUTTON_LINK_URL}],
-            timing: {showOnStart: true}
-          }
-        ]
-      },
-      [TITLE, BUTTON_1_LABEL]
-    );
-  });
-  it('should show title and two buttons', () => {
-    expectContainsInOverlay(
-      {
-        title: 'cta title',
-        messages: [
-          {
-            title: 'cta title',
-            buttons: [
-              {label: BUTTON_1_LABEL, link: BUTTON_LINK_URL},
-              {label: BUTTON_2_LABEL, link: BUTTON_LINK_ENTRY}
-            ],
-            timing: {showOnStart: true}
-          }
-        ]
-      },
-      [TITLE, BUTTON_1_LABEL, BUTTON_2_LABEL]
-    );
-  });
-  it('should show description and one button', () => {
-    expectContainsInOverlay(
-      {
-        messages: [
-          {
-            description: DESCRIPTION,
-            buttons: [{label: BUTTON_1_LABEL, link: BUTTON_LINK_ENTRY}],
-            timing: {showOnStart: true}
-          }
-        ]
-      },
-      [DESCRIPTION, BUTTON_1_LABEL]
-    );
-  });
-  it('should show description and two buttons', () => {
-    expectContainsInOverlay(
-      {
-        description: 'cta description',
-        messages: [
-          {
-            description: DESCRIPTION,
-            buttons: [
-              {label: BUTTON_1_LABEL, link: BUTTON_LINK_URL},
-              {label: BUTTON_2_LABEL, link: BUTTON_LINK_ENTRY}
-            ],
-            timing: {showOnStart: true}
-          }
-        ]
-      },
-      [DESCRIPTION, BUTTON_1_LABEL, BUTTON_2_LABEL]
-    );
-  });
-  it('should show title, description and one button', () => {
-    expectContainsInOverlay(
-      {
-        messages: [
-          {
-            title: TITLE,
-            description: DESCRIPTION,
-            buttons: [{label: BUTTON_1_LABEL, link: BUTTON_LINK_URL}],
-            timing: {showOnStart: true}
-          }
-        ]
-      },
-      [TITLE, DESCRIPTION, BUTTON_1_LABEL]
-    );
-  });
-  it('should show title, description and two buttons', () => {
-    expectContainsInOverlay(
-      {
-        messages: [
-          {
-            title: TITLE,
-            description: DESCRIPTION,
-
-            buttons: [
-              {label: BUTTON_1_LABEL, link: BUTTON_LINK_URL},
-              {label: BUTTON_2_LABEL, link: BUTTON_LINK_ENTRY}
-            ],
-            timing: {showOnStart: true}
-          }
-        ]
-      },
-      [TITLE, DESCRIPTION, BUTTON_1_LABEL, BUTTON_2_LABEL]
-    );
-  });
-  describe('buttons', () => {
-    describe('close button', () => {
-      it('should close popup on click', () => {
-        expectCloseButtonInOverlay({
+  describe('overlay content', () => {
+    it('should show title', () => {
+      expectContainsInOverlay({messages: [{title: TITLE, timing: {showOnStart: true}}]}, [TITLE]);
+    });
+    it('should show description', () => {
+      expectContainsInOverlay({messages: [{description: DESCRIPTION, timing: {showOnStart: true}}]}, [DESCRIPTION]);
+    });
+    it('should show one button', () => {
+      expectContainsInOverlay(
+        {
+          messages: [{buttons: [{label: BUTTON_1_LABEL, link: BUTTON_LINK_URL}], timing: {showOnStart: true}}]
+        },
+        [BUTTON_1_LABEL]
+      );
+    });
+    it('should show two buttons', () => {
+      expectContainsInOverlay(
+        {
           messages: [
             {
+              buttons: [
+                {label: BUTTON_1_LABEL, link: BUTTON_LINK_URL},
+                {label: BUTTON_2_LABEL, link: BUTTON_LINK_ENTRY}
+              ],
+              timing: {showOnStart: true}
+            }
+          ]
+        },
+        [BUTTON_1_LABEL, BUTTON_2_LABEL]
+      );
+    });
+    it('should show title and description', () => {
+      expectContainsInOverlay(
+        {
+          messages: [{title: TITLE, description: DESCRIPTION, timing: {showOnStart: true}}]
+        },
+        [TITLE, DESCRIPTION]
+      );
+    });
+    it('should show title and one button', () => {
+      expectContainsInOverlay(
+        {
+          messages: [
+            {
+              title: TITLE,
               buttons: [{label: BUTTON_1_LABEL, link: BUTTON_LINK_URL}],
               timing: {showOnStart: true}
             }
           ]
+        },
+        [TITLE, BUTTON_1_LABEL]
+      );
+    });
+    it('should show title and two buttons', () => {
+      expectContainsInOverlay(
+        {
+          title: 'cta title',
+          messages: [
+            {
+              title: 'cta title',
+              buttons: [
+                {label: BUTTON_1_LABEL, link: BUTTON_LINK_URL},
+                {label: BUTTON_2_LABEL, link: BUTTON_LINK_ENTRY}
+              ],
+              timing: {showOnStart: true}
+            }
+          ]
+        },
+        [TITLE, BUTTON_1_LABEL, BUTTON_2_LABEL]
+      );
+    });
+    it('should show description and one button', () => {
+      expectContainsInOverlay(
+        {
+          messages: [
+            {
+              description: DESCRIPTION,
+              buttons: [{label: BUTTON_1_LABEL, link: BUTTON_LINK_ENTRY}],
+              timing: {showOnStart: true}
+            }
+          ]
+        },
+        [DESCRIPTION, BUTTON_1_LABEL]
+      );
+    });
+    it('should show description and two buttons', () => {
+      expectContainsInOverlay(
+        {
+          description: 'cta description',
+          messages: [
+            {
+              description: DESCRIPTION,
+              buttons: [
+                {label: BUTTON_1_LABEL, link: BUTTON_LINK_URL},
+                {label: BUTTON_2_LABEL, link: BUTTON_LINK_ENTRY}
+              ],
+              timing: {showOnStart: true}
+            }
+          ]
+        },
+        [DESCRIPTION, BUTTON_1_LABEL, BUTTON_2_LABEL]
+      );
+    });
+    it('should show title, description and one button', () => {
+      expectContainsInOverlay(
+        {
+          messages: [
+            {
+              title: TITLE,
+              description: DESCRIPTION,
+              buttons: [{label: BUTTON_1_LABEL, link: BUTTON_LINK_URL}],
+              timing: {showOnStart: true}
+            }
+          ]
+        },
+        [TITLE, DESCRIPTION, BUTTON_1_LABEL]
+      );
+    });
+    it('should show title, description and two buttons', () => {
+      expectContainsInOverlay(
+        {
+          messages: [
+            {
+              title: TITLE,
+              description: DESCRIPTION,
+
+              buttons: [
+                {label: BUTTON_1_LABEL, link: BUTTON_LINK_URL},
+                {label: BUTTON_2_LABEL, link: BUTTON_LINK_ENTRY}
+              ],
+              timing: {showOnStart: true}
+            }
+          ]
+        },
+        [TITLE, DESCRIPTION, BUTTON_1_LABEL, BUTTON_2_LABEL]
+      );
+    });
+    describe('buttons', () => {
+      describe('close button', () => {
+        it('should close popup on click', () => {
+          expectCloseButtonInOverlay({
+            messages: [
+              {
+                buttons: [{label: BUTTON_1_LABEL, link: BUTTON_LINK_URL}],
+                timing: {showOnStart: true}
+              }
+            ]
+          });
         });
       });
-    });
-    describe('one button', () => {
-      it('should open new window if link is a url', () => {
-        expectWindowOpenInOverlay(
-          {
-            messages: [
-              {
-                buttons: [{label: BUTTON_1_LABEL, link: BUTTON_LINK_URL}, {}],
-                timing: {showOnStart: true}
-              }
-            ]
-          },
-          BUTTON_1_LABEL,
-          BUTTON_LINK_URL
-        );
-      });
-      it('should call loadMedia if link is not a url', () => {
-        expectLoadMediaInOverlay(
-          {
-            messages: [
-              {
-                buttons: [{label: BUTTON_1_LABEL, link: BUTTON_LINK_ENTRY}],
-                timing: {showOnStart: true}
-              }
-            ]
-          },
-          BUTTON_1_LABEL,
-          BUTTON_LINK_ENTRY
-        );
-      });
-    });
-    describe('two buttons', () => {
-      describe('click on button 1', () => {
+      describe('one button', () => {
         it('should open new window if link is a url', () => {
           expectWindowOpenInOverlay(
             {
               messages: [
                 {
-                  buttons: [
-                    {label: BUTTON_1_LABEL, link: BUTTON_LINK_URL},
-                    {label: BUTTON_2_LABEL, link: BUTTON_LINK_ENTRY}
-                  ],
+                  buttons: [{label: BUTTON_1_LABEL, link: BUTTON_LINK_URL}, {}],
                   timing: {showOnStart: true}
                 }
               ]
@@ -226,10 +477,7 @@ describe('call to action overlay', () => {
             {
               messages: [
                 {
-                  buttons: [
-                    {label: BUTTON_1_LABEL, link: BUTTON_LINK_ENTRY},
-                    {label: BUTTON_2_LABEL, link: BUTTON_LINK_URL}
-                  ],
+                  buttons: [{label: BUTTON_1_LABEL, link: BUTTON_LINK_ENTRY}],
                   timing: {showOnStart: true}
                 }
               ]
@@ -239,40 +487,78 @@ describe('call to action overlay', () => {
           );
         });
       });
-      describe('click on button 2', () => {
-        it('should open new window if link is a url', () => {
-          expectWindowOpenInOverlay(
-            {
-              messages: [
-                {
-                  buttons: [
-                    {label: BUTTON_1_LABEL, link: BUTTON_LINK_ENTRY},
-                    {label: BUTTON_2_LABEL, link: BUTTON_LINK_URL}
-                  ],
-                  timing: {showOnStart: true}
-                }
-              ]
-            },
-            BUTTON_2_LABEL,
-            BUTTON_LINK_URL
-          );
+      describe('two buttons', () => {
+        describe('click on button 1', () => {
+          it('should open new window if link is a url', () => {
+            expectWindowOpenInOverlay(
+              {
+                messages: [
+                  {
+                    buttons: [
+                      {label: BUTTON_1_LABEL, link: BUTTON_LINK_URL},
+                      {label: BUTTON_2_LABEL, link: BUTTON_LINK_ENTRY}
+                    ],
+                    timing: {showOnStart: true}
+                  }
+                ]
+              },
+              BUTTON_1_LABEL,
+              BUTTON_LINK_URL
+            );
+          });
+          it('should call loadMedia if link is not a url', () => {
+            expectLoadMediaInOverlay(
+              {
+                messages: [
+                  {
+                    buttons: [
+                      {label: BUTTON_1_LABEL, link: BUTTON_LINK_ENTRY},
+                      {label: BUTTON_2_LABEL, link: BUTTON_LINK_URL}
+                    ],
+                    timing: {showOnStart: true}
+                  }
+                ]
+              },
+              BUTTON_1_LABEL,
+              BUTTON_LINK_ENTRY
+            );
+          });
         });
-        it('should call loadMedia if link is not a url', () => {
-          expectLoadMediaInOverlay(
-            {
-              messages: [
-                {
-                  buttons: [
-                    {label: BUTTON_1_LABEL, link: BUTTON_LINK_URL},
-                    {label: BUTTON_2_LABEL, link: BUTTON_LINK_ENTRY}
-                  ],
-                  timing: {showOnStart: true}
-                }
-              ]
-            },
-            BUTTON_2_LABEL,
-            BUTTON_LINK_ENTRY
-          );
+        describe('click on button 2', () => {
+          it('should open new window if link is a url', () => {
+            expectWindowOpenInOverlay(
+              {
+                messages: [
+                  {
+                    buttons: [
+                      {label: BUTTON_1_LABEL, link: BUTTON_LINK_ENTRY},
+                      {label: BUTTON_2_LABEL, link: BUTTON_LINK_URL}
+                    ],
+                    timing: {showOnStart: true}
+                  }
+                ]
+              },
+              BUTTON_2_LABEL,
+              BUTTON_LINK_URL
+            );
+          });
+          it('should call loadMedia if link is not a url', () => {
+            expectLoadMediaInOverlay(
+              {
+                messages: [
+                  {
+                    buttons: [
+                      {label: BUTTON_1_LABEL, link: BUTTON_LINK_URL},
+                      {label: BUTTON_2_LABEL, link: BUTTON_LINK_ENTRY}
+                    ],
+                    timing: {showOnStart: true}
+                  }
+                ]
+              },
+              BUTTON_2_LABEL,
+              BUTTON_LINK_ENTRY
+            );
+          });
         });
       });
     });
